requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==22.12.1
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
confusable-homoglyphs==3.2.0
constantly==15.1.0
cryptography==39.0.0
daphne==4.0.0
defusedxml==0.7.1
Django==3.2.16
django-allauth==0.52.0
django-registration==3.3
djangorestframework==3.14.0
docker==6.0.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.0.0
incremental==22.10.0
msgpack==1.0.4
oauthlib==3.2.2
packaging==23.0
Pillow==9.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.7
pywin32==305
redis==4.4.1
requests==2.28.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.13
websocket-client==1.4.2
zipp==3.11.0
zope.interface==5.5.2
